refactor: refactor repositories to improve product association and update queries for trending products
